Add --queue-bypass to setting up iptable rules
I found this critically important b/c if nothing is listening to the queue (you have started suricata yet, or stopped it), packets are simply stuck and the server becomes unusable on existing and new connections, and requires a reboot to flush iptables.  If there is a better way of handling this, please replace with that.  
Thanks.