unix suricata-command.socket permissions set
So far, the suricata socket suricata-command.socket has the rights rw-r----- suricata:user.
When suricata is used with restricted access, an other application (suricatasc) that needs to access to the command socket
also with restricted access can not write to the socket since it is not the owner
(e.g suricata within an LXC container, with an hardened value for umask and hardened rights for users).
The socket should be set as rw-rw----.
Use chmod instead of fchmod and set it after the socket creation.