NSM: provide filemane and MD5 for each attachment record within SMTP flow
The goal is to add extra information for network security monitoring: so far it logs only MD5 body and subject.
I propose to log MD5 attached filename detected in SMTP protocol within json record.
This is proposed as an option.
It can be activated by using the attachments parameter (smtp:md5:attachments).