Adds PID to file name, writes to a temp file.
Adds the pid to the file name to reduce the likelihood that a file is
overwritten when suricata is unexpectedly killed.  The number in the
waldo file is only written out during a clean shutdown.  In the event
of an improper shutdown, extracted files will be written using the old
number and existing files with the same name will be overwritten.

Writes extracted files and their metadata to a temporary file prefixed
with a '.'.  Renames the files when they are completely done being
written and the meta file is closed.  As-is there is no way to know that
a file on disk is still being written to by suricata.