Adds options to mark when a file is final.
This takes the form of an option to have a working file prefix, which is
removed when files are finalized, and an option to add the pid of the
process to file names.

Adding the pid to the file name reduces the likelihood that a file is
overwritten when suricata is unexpectedly killed. The number in the
waldo file is only written out during a clean shutdown. In the event
of an improper shutdown, extracted files will be written using the old
number and existing files with the same name will be overwritten.

Writes extracted files and their metadata to a temporary file prefixed
with a '.'. Renames the files when they are completely done being
written and the meta file is closed. As-is there is no way to know that
a file on disk is still being written to by suricata.