Rust: add a common interface to declare and register all parsers
Add a common structure definition to contain all required elements
required for registering a parser.
This also reduces the risk of incorrectly registering a parser: the
compiler prevents omitting required functions from the structure, and
functions (even if external) are type-checked. Optional functions are
explicitly marked.

The registration function has the same behavior than the C usual
registration code, it just calls the C functions directly. Using the
Parser structure, it allows factorization of the code.

If extra registration code (for functions not in the structure, for ex.
Files extraction), it is still possible by calling the C functions after
the registration.