app-layer: use FILE_HAS_GAPS flag in http & smtp
File structure has the possibility to have FILE_HAS_GAPS flag. This
will cause eve-log to dump a field "gaps: <TRUE/FALSE>" in fileinfo.
This flag is currently used only in RUST parser.

This commit will allow http and smtp parsers to set the FILE_HAS_GAPS
flag, using the Truncate handler.
There is a caveat. The Truncation of the file will set the state to
FILE_STATE_TRUNCATED. However, this will stop the detection phase,
possibly even before its first run. This is solved by preventing setting
the state of the file to FILE_STATE_TRUCATED in FileCloseFilePtr. In
any case, the state will be updated in the output-data (or
output-filedata) at the end of the detection phase.