counter: Counter reset and max_parallel_flows
This feature adds the ability to reset the value of a counter. This
feature is controlled by the flag STATS_FLAGS_RESETTING. This flag will
trigger a reset (via function StatsReset) of the value of the counter
at the end of function StatsOutput and StatsOutputCounterSocket.

An example counter is provided. This counter counts the maximum number
of parallel flows at any single time, increasing and decreasing an
atomic variable every time we create/destroy a flow.