stream-tcp-reassemble: use the correct stream_depth
When a new tcp session is created, the reassemby_depth
value is set by stream_config.reassembly_depth but
it could happen that then this value is updated.
For instance, when file-store keyword is used
the function TcpSessionSetReassemblyDepth is called
to update the value.

StreamTcpReassembleCheckDepth doesn't use the value updated
which is represented by ssn->reassembly_depth, but use the old
one represented by stream_config.reassembly_depth.

As a result, reassembly_depth value pointed by ssn is not
considered when a stream is reassembled, and a stream can be
truncated because the wrong reassembly_depth value is used.

To fix this, TcpSessionSetReassemblyDepth now checks on
ssn->reassembly_depth and not on stream_config.reassembly_depth.

Issue #2264