app-layer-htp: use file-store.stream-depth
If a signature contains the filestore keyword,
a file is truncated if its size reaches a limit
such as request/response body limit even when
when file-store.stream-depth is set because it
is not evaluated, then the files are truncated
as a conseguence.
This adds a flag for stream depth that is set
when the filestore keyword is fired and in that
case the content length will be checked against
stream depth, otherwise libhtp's limits are considered.