startup: Check if logrotation will still work after priv drop (#2386)
Write permission on the logdir is necesarry to create new files in there,
which is necesary after logrotation to continue logging. (use cases:
exe.json and others)

If suricata could open the log files but not create new ones it silently failed,
not writing any logs after the first logrotation.
This patch adds a check before priv drop and warns+exit()s
if we will not be able to create new files after priv drop.

If we won't drop privs we check if the current (and general) user will be able
to create logfiles after log rotation.

For details and discussion see:
https://redmine.openinfosecfoundation.org/issues/2386