suricata: add display-memcaps command
When global memcap is enabled and it is reached
by all the other memcaps, suricata won't start.
For a user it could be helpful to have a simple
way to display all the memcaps value so the values
can be easily checked.
This is done by the new command 'display-memcaps'
added by this patch.