fileinfo: fix IP parameters order in events
All fileinfo events should have a source IP that is the one sending
the data and a destination IP that is the receiving part. This was
not the case as the side information was not contained in the File
structure.

This patch adds a side information to the File allowing the
application layer to define which side should be considered at
logging time.