Restore core dump privileges after uid/gid change
Per man(2) prctl, this flag is set to the value of '/proc/sys/fs/suid_dumpable' - which defaults to zero - when the effective user and/or group for the process is changed, effectively preventing suricata from creating core dumps after the main thread drops capabilities and changes the process uid/gid. This PR resets the flag (in order to allow suricata to create core dumps) after the process's effective UID/GID have been changed.