configure: handle --disable-lua/--disable-luajit correctly
For both lua and luajit, current version of configure.ac enables these
features in the if-present action of respective AC_ARG_ENABLE
declarations. This result in them being enabled even if the relevant
command-line arguments were in fact *--disable-foo*; the only way of
disabling them is to omit the arguments from the command line
altogether.

Make the if-present action for both arguments null so that configure
respects both positive and negative command-line options if provided.
Both if-absent actions remain as they were so that lua/luajit support is
disabled by default.